refactor application of fiberNorms
We measure the fiberNorms from the pfsArm files, so we don't want the
pfsArm files to include the fiberNorms (or it would be difficult to keep
track of what everything was relative to). Instead, the fiberNorms are
applied in mergeArms (before sky subtraction) and also in coaddSpectra,
using a new function to make the application. As part of this, the screen
response has moved to its own file, so it can be easily updated.

Note that the photometry (measured in reduceExposure) is no longer
corrected for the fiberNorms. This probably wouldn't be difficult to do,
but it wasn't clear we needed it.