move screen response correction
I had put the screen response correction alongside the application
of the fiberNorms. However, we don't want the screen response
included in the fiberNorms, so we need to remove it before they
are measured. The screen response is now a Task that's called by
ReduceExposureTask; it modifies the spectrum before the fiberNorms
are measured.
Also realised that the normalisation spectrum in the fiberProfiles
needs to be corrected for the screen response, so added the screen
response to measureFiberNorms.