add barycentric correction
It's applied in reduceExposure, so that it doesn't need to be repeated
(although it will be repeated for each arm separately, but it will then
get baked into the pfsArm spectra, so we don't need to repeat the
correction in both mergeArms and coaddSpectra). The correction values are
recorded in the pfsArm.notes.

The implementation follows the helpful example from astropy:
https://docs.astropy.org/en/stable/coordinates/velocities.html#id5