be more careful about masks in the extraction
We want two different ways of setting the mask for a spectral
pixel. If the extraction is successful, we want to set the mask
to be the OR of contributing pixels satisfying minFracMask. If
the extraction is not successful, we want to set the mask to
be the OR of pixels satisfying minFracMask that did NOT
contribute, and NO_DATA (and BAD_FIBERTRACE if that's appropriate).
This way, we can identify bad spectral pixels (NO_DATA) as well
as the reason they're bad (BAD, CR, etc.).