refactor barycentric correction
I put the barycentric correction into reduceExposure, producing barycentric
corrected pfsArm files, figuring that it would save the complication of
applying it twice (once in mergeArms and later in coaddSpectra). However,
we need uncorrected wavelengths in mergeArms for the 1d sky subtraction.
To deal with this, split the barycentric correction into calculation and
application. The calculation is done in reduceExposure, and then it's
applied in mergeArms (after the sky subtraction) and coaddSpectra.