add certifyDetectorMaps and implement detectorMap naming scheme
Building PFS detectorMaps frequently uses the calib detectorMap as part
of the process of fitting a new detectorMap (e.g., for the centroiding
starting point, and the slit offsets). In order to avoid cycles when
building the detectorMap, we write the new detectorMap as a separate
dataset type from the calib detectorMap dataset type. This new
function+command-line copies the new detectorMaps as calib detectorMaps,
and then certifies them for use as calibs.

As part of this, I've implemented an RHL-approved naming scheme for the
detectorMap products:
* detectorMap_bootstrap is the output from bootstrapping, i.e., close but
  non-precise.
* detectorMap_candidate is the output from fitting a new detectorMap from
  arcs.
* detectorMap is the output from adjusting a detectorMap when processing
  a science exposure. This is our best guess of the mapping of
  fiberId,wavelength to x,y for a particular exposure.
* detectorMap_calib is the certified calib used for processing (precise,
  but slightly inaccurate for any particular exposures other than the arcs
  they were fit from).