add infrastructure for using fiberNorms
The fiberNorms_calib have unusual dimensions, so the Gen3 middleware needs
some help to find the correct ones. Added a lookupFunction that finds the
fiberNorms for an exposure by iterating over arms, and added an extra
task prior to the coadd that records which fiberNorms to use for an
exposure (since that's not available in the dataId).