Spectrum: remove background, covariance, spectrum
Spectrum is essentially a carrier for the results from
FiberTraceSet::extractSpectra. I'm now finally getting rid of the things
that have sat in Spectrum for a long time that we never use.

The background is not measured as part of the extraction, so there's no
need for it.

The covariance that we've been putting in is not the covariance that was
originally intended. I intend to change how the variance is calculated and
not bother with the covariance, so let's rip it out.
with the

"spectrum" has been a synonym for flux for a long time. It was never a good
name, so let's finally let it die. I was surprised how many places it has
hung around.