extractSpectra: reduce number of masked and zero pixels
Changed the logic of how the final spectral pixels are set, so that we
always set a value if it's available. Changed the calculation of the
variance, since we don't really care about the covariance, and it's much
simpler.
We can sometimes get fibers for which we have only a single pixel
contributing to the matrix, and it's well out in the wings. To avoid
getting garbage values, mask a pixel as SUSPECT (defined by LSST's Mask)
if we don't reach a threshold (called 'minFracImage', defaulting to ~ 0.3)
of relative contributions from unmasked pixels.