refactor ModelBasedDetectorMap slightly
to separate model evaluation and findPoint, make sampling clearer and
more complete.

Separating model evaluation (now in the evalModel method) from findPoint
allows subclasses to implement findPoint differently (and perhaps faster
than model evaluation, using the splines).

The sampling is clearer by requiring the subclass to provide the
dispersion (nm/pixel) and then specifying the sampling frequency in pixels.
I've also added a precision parameter, which controls how close to the
edge we can get.

The sampling is now done with more care: we find a position for the fiber
that's on the detector (rather than assuming that the center will work),
and as we approach the edge we reduce the step size, attempting to get
points as close to the edge as possible.