optimize FiberTraceSet::extractSpectra
Changes made as part of PIPE2D-1507 (reduce number of masked
and zero pixels) resulted in the spectral extraction taking a few
seconds longer (3.75 sec in w.2024.22 --> 6.35 sec, about a 70%
increase). Tracked this down to the summing of the trace while
respecting the mask: the mask had not been respected in w.2024.22,
but it used Eigen expressions to do the sum. I discovered how to
implement the new sum using Eigen expressions, and this has
reduced the time for the spectral extraction back to the original
level (the summing of the trace takes a bit longer, but the matrix
construction and solution is faster, so it's a wash: 3.69 sec now).