fitFocalPlane: identify and remove fully-masked fibers
Sometimes a fiber's spectrum is completely masked (e.g., when the
amplifier is having problems). In that case, remove those fibers
from the fit. This keeps us breaking when we require that a
bunch of sequential fibers have some good spectra (e.g., for
sky subtraction, with the block approach).