reduceProfiles: convert to Gen3
The Gen3 process for generating fiberProfiles doesn't involve adding the
normalization, since it is a different set of exposures that need to be
processed separately. I was trying to avoid a two-step process, and we've
talked about stripping the normalizations out of the fiberProfiles anyway
(they will go in the fiberNorms).

In the process, renamed the fiberProfiles.yaml pipeline definition to
measureFiberProfiles.yaml: we're going to distinguish between measuring
profiles (Gen2 equivalent: constructFiberProfiles) and fitting profiles
(Gen2 equivalent: reduceProfiles). Also, the merging process is different
so we need to have different names for those: measure+combine and
fit+merge.