refactor coaddSpectra,fitFluxCal to use batched outputs again
This effectively reverts the change made in PIPE2D-1525 to have these
produce one file per object. This was very inefficient, so that even simple
pipelines wouldn't even build within half an hour.

Instead, we batch the outputs by catId (we could imagine doing this by
catId,tract,patch later if necessary). Further, we use the new
"combination" dimension that defines exposures that contribute to the
coadd, which allows us to create different combinations for the same
spectra in the same collection.