measureFiberNorms: remove pfsConfig
Because MeasureFiberNormsConnections operates over arm (and not exposure),
the pfsConfig connection was providing all possible pfsConfigs in the
registry. It's not clear how to deal with that without some dirty hacking
(a lookupFunction would not know which exposures to select, and
adjustQuantum runs after it has already selected everything). However,
we were only using the pfsConfig for plotting, and that has been relegated
to the QA, so we may as well strip it out.