mergeFiberProfiles: fix selection of pfsConfig
Because MergeFiberProfilesConnections doesn't operate over exposure,
the pfsConfig connection was providing all possible pfsConfigs in the
registry (this would bring the machine to its knees when running 12
merges in parallel). Added a lookupFunction to get the list of exposures
from the profiles_run, and then use that to provide only the pfsConfigs
we need.