reduceExposure: fix regression in screen correction
The conversion of spectra to pfsArm used to happen after the
screen response was applied. When it moved to earlier, I
neglected to change the screen correction to be applied to
the pfsArm rather than spectra. Therefore, the screen
correction was being applied, but the result wasn't being
seen in the pfsArm. Fixed by correcting the pfsArm.
This involved upgrading the screen correction to operate
on a PfsFiberArraySet rather than SpectrumSet.