Make TransmissionCurve.photometer() return stochastic error
An option is added to TransmissionCurve.photometer() to return
stochastic error as well as the integrated flux.

The option `quadpack` is removed by this commit, and quadpack is
no longer used. The reason of this removal is that the stochastic
error of the integral is well defined only after an approximation
formula for the integral is fixed. Because only flux calibration
tasks use TransmissionCurve.photometer(), the effect of this removal
is limited.