OversampledSpline: use median x instead of center of bin
Sometimes, samples can pile up on one side of the bin, and when
that happens on the edge of a line, the robust spline can be
biased, leading to unnecessary rejection of good points. Using
the median x position instead of the center of the bin results
in a better representation of the data.