reduceExposure: restore requireAdjustDetectorMap
The requireAdjustDetectorMap configuration parameter (and
associated feature) was removed in the Gen3 transition
(in the process of simplifying reduceExposure). This was
a mistake: we still want this feature. Reconstructed the
feature. It's currently set to False (as it was before),
but I'm setting it to True for pipelines I think require
it (fiberNorms, calibrateExposure/science).