fitFluxCal: Calibrate spectra before interpolation
Previously, we interpolated those parts of a spectrum that are missing
due to absent arms, then in the optimization loop we divided it by the
calibration vector (to be optimized), and then we integrated the
spectrum to compute broadband fluxes. This order is reproved because
the spectrum before flux calibration is so much affected by telluric
absorption and dichroics that no simple interpolation looks good.

We therefore change the order: divide the spectrum by a rough
calibration vector, and then interpolate, before entering the
optimization loop.