fitFluxCal: Accept flux standards from HSC and Gaia
Previously, only PS1 was available as the source of flux standards.
It was because we interpolate in a complicated way those parts of
spectra that are missing due to absent arms. This interpolation
routine was valid only with PS1 fluxes. We now accept flux standards
from HSC and Gaia by extending the interpolation routine.