lsf: change how LSFs are warped, coadded and convolved
The m arm has a very different LSF than b and n, and we're switching to
a wavelength sampling scheme for merged spectra that will make that even
more different. We want to be able to use the correct LSF according to the
the part of the spectrum we're interested in. Making this possible involves
several coordinated changes to LSF warping, coaddition and convolution.

Removed coaddLsf function in favor of a new CoaddLsf class. CoaddLsf is
similar to LSST's CoaddPsf: it has a list of component LSFs, and combines
the appropriate ones when asked for the LSF. This will allow us to use the
warped b LSF when we look at the b part of a merged spectrum, etc.

Converted the warpPsf function to a method on the Lsf class, because we
want to treat warping differently according to what kind of Lsf we have.
We have implementations for the GaussianLsf and CoaddLsf classes, but none
for the FixedEmpiricalLsf and ExtractionLsf classes (but we're not using
those at the moment).

Added a convolve method on the Lsf class, since we want to treat
convolution different according to the kind of LSF (and it makes it easier
for the user: now only one call is required to convolve by the LSF).

In support of this, added methods to copy and add kernels.

Modified code that uses these facilities and added tests.

To help with this, added methods to add kernels, including
GaussianKernel1D, which will allow us to combine those.