add SubtractSky1dTask, deprecate subtractSky1d
Since the fiber throughputs seem to vary from what is expected
from the fiberNorms, SubtractSky1dTask offers the option to
fit a scaling factor for each fiber, so the sky lines from
the model match the data. Currently this is a single scale
factor per fiber, but this technique could be expanded to
allow a scaling as a function of wavelength.

Deprecated subtractSky1d function to encourage people to use
the better SubtractSky1dTask.