calibs: work around DM-47920
cp_pipe is writing HIERARCH keys including a period are
triggering a bug in reading WCS. This bug was fixed in
afw a couple of months after the 28.0.1 tag. I don't want
to go to the trouble of figuring out how to patch afw,
so I'm patching cp_pipe's CalibCombineTask so that it
doesn't trigger the bug.