move calculation of FiberTrace positions to DetectorMap
DetectorMap knows where the chip gap is, while FiberProfile only
knows the shape of the fiber, so we need DetectorMap to tell
FiberProfile where the trace is. Since a trace can be slightly
off the detector, with the outer parts of the profile still on
the detector, we need DetectorMap to tell FiberProfile not just
the center of the trace but all the pixels of interest.