expand 1d sky model
The 1d sky model now consists of three components: the blocked, oversampled
splines (as currently) for the sky as a function of wavelength (and some
variation over the detector); "skyNorms", which is an additional
normalization to apply to the spectrum from the splines, measured from the
sky lines of many exposures; and a 2D polynomial over the focal plane that
allows for additional variation of the normalization.

This commit moves the construction of the 1d sky model into the new
FitSky1dTask, and adds the MeasureSkyNormsTask and related skyNorms.yaml
pipeline for building the skyNorms.