perform coadds over groups of objects within a catId
pfsCoadds are getting too large, so we're subdividing a catId
into groups of objects (dimension name obj_group). The mapping
between objId and objGroup is provided by the ObjectGroupMap,
which we determine when we define a combination, and which is
written (to PFS/objectGroups) for later use. This allows us to
select only the objId in the objGroup of interest when we coadd.