use adjust_all_quanta to consolidate cosmic-ray tasks
The CompareCosmicRayTask requires grouping of visits in the
database which is a burden (especially when observing). The new
(to us) PipelineTaskConnections.adjust_all_quanta functionality
allows us to group visits dynamically at pipeline build time.
This further allows us to consolidate the CosmicRayTask and the
CompareCosmicRayTask, with the previous difference now achieved
by different grouping.