fitDetectorMap: gather slit offsets before fitting
The fitDetectorMap task's slitOffsets connection was not related
to any visit, and therefore the lookup function returns all
available detectorMaps, resulting in a 'ScalarError: Found
multiple datasets'. The slitOffsets need to be related to a
visit, so added a new task that gets the slit offsets for
each visit. Then we make sure they're all consistent.

Introduced the 'doSlitOffsets' pipeline parameter to keep
the value in the new task and the fitting task coordinated.