NormalizedPolynomialND: change normalization scheme
The code has been normalizing inputs to the range 0..1, but this provides
a much-reduced function space. Normalizing to -1..1, as is done in other
normalized polynomial classes, is better as it provides both odd and even
functions.

We have to support the old normalization scheme, since there's existing
data we need to be able to read and use, so the normalization scheme can
be chosen at construction time.