add DetermineFluxReferenceTask
This derivative of FitFluxReferenceTask is intended to provide
a quick and dirty flux reference using the Gaia catalog, for
real-time observing support. It is capable of using a local
Gaia catalog or failing over to remote lookup. The former
takes about 40 sec per visit (with the time dominated by the
spectral integration) and the latter about 90 sec per visit.