add coaddNoise task
This task does a Monte Carlo procedure on the coadd, allowing
measurement of actual variance and covariance. We set all input
pfsArms to a known flux, add random noise according to the
variance array, and coadd; this process is repeated a configurable
number of times. From the ensemble of such coadds, we write an
output that looks like a pfsObject, but the meaning of the
members is different: the flux is the mean bias, the
variance/covariance are measured from the samples, and the sky
is a no-noise combined spectrum.